게임 시작시 로딩이 느려지는 현상 일시적 해결

+++ Nolgilo/Classes/DBcore.h
@@ -32,5 +32,5 @@
 -(NSArray *) RoomListData;
 -(void)RoomConnectUpdate:(int)roomid:(int)connuser;
 -(int)RoomConnectUser:(int)roomid;
--(void)RoomClose:(NSInteger)roomid;
+-(void)RoomClose:(int)roomid;
 @end
